+++ docs/spec/src/index.ts
@@ -1,9 +1,16 @@
 import { IntegrationSpecConfig } from '@jupiterone/integration-sdk-core';
 
 import { IntegrationConfig } from '../../../src/config';
-import { accessSpec } from './access';
 import { accountSpec } from './account';
+import { applicationsSpec } from './application';
+import { projectsSpec } from './projects';
+import { usersSpec } from './users';
 
 export const invocationConfig: IntegrationSpecConfig<IntegrationConfig> = {
-  integrationSteps: [...accountSpec, ...accessSpec],
+  integrationSteps: [
+    ...accountSpec,
+    ...projectsSpec,
+    ...usersSpec,
+    ...applicationsSpec,
+  ],
 };