+++ src/config.ts
@@ -4,46 +4,17 @@ import {
   IntegrationInstanceConfigFieldMap,
   IntegrationInstanceConfig,
 } from '@jupiterone/integration-sdk-core';
-import { createAPIClient } from './client';
+import { Client } from './google-cloud/client';
+import { parseServiceAccountKeyFile } from './utils/parseServiceAccountKeyFile';
 
-/**
- * A type describing the configuration fields required to execute the
- * integration for a specific account in the data provider.
- *
- * When executing the integration in a development environment, these values may
- * be provided in a `.env` file with environment variables. For example:
- *
- * - `CLIENT_ID=123` becomes `instance.config.clientId = '123'`
- * - `CLIENT_SECRET=abc` becomes `instance.config.clientSecret = 'abc'`
- *
- * Environment variables are NOT used when the integration is executing in a
- * managed environment. For example, in JupiterOne, users configure
- * `instance.config` in a UI.
- */
 export const instanceConfigFields: IntegrationInstanceConfigFieldMap = {
-  clientId: {
+  serviceAccountKeyFile: {
     type: 'string',
-  },
-  clientSecret: {
-    type: 'string',
-    mask: true,
   },
 };
 
-/**
- * Properties provided by the `IntegrationInstance.config`. This reflects the
- * same properties defined by `instanceConfigFields`.
- */
 export interface IntegrationConfig extends IntegrationInstanceConfig {
-  /**
-   * The provider API client ID used to authenticate requests.
-   */
-  clientId: string;
-
-  /**
-   * The provider API client secret used to authenticate requests.
-   */
-  clientSecret: string;
+  serviceAccountKeyFile: string;
 }
 
 export async function validateInvocation(
@@ -51,12 +22,17 @@ export async function validateInvocation(
 ) {
   const { config } = context.instance;
 
-  if (!config.clientId || !config.clientSecret) {
+  if (!config.serviceAccountKeyFile) {
     throw new IntegrationValidationError(
-      'Config requires all of {clientId, clientSecret}',
+      'Config requires all of { serviceAccountKeyFile }',
     );
   }
 
-  const apiClient = createAPIClient(config);
+  const apiClient = new Client({
+    serviceAccountKeyFile: config.serviceAccountKeyFile,
+    serviceAccountKeyConfig: parseServiceAccountKeyFile(
+      config.serviceAccountKeyFile,
+    ),
+  });
   await apiClient.verifyAuthentication();
 }