+++ src/index.ts
@@ -1,5 +1,7 @@
 import { IntegrationInvocationConfig } from '@jupiterone/integration-sdk-core';
 import { integrationSteps } from './steps';
+import { deserializeIntegrationConfig } from './utils/integrationConfig';
+import getStepStartStates from './getStepStartStates';
 import {
   validateInvocation,
   IntegrationConfig,
@@ -9,6 +11,9 @@ import {
 export const invocationConfig: IntegrationInvocationConfig<IntegrationConfig> =
   {
     instanceConfigFields,
+    getStepStartStates,
     validateInvocation,
     integrationSteps,
   };
+
+export { IntegrationConfig, deserializeIntegrationConfig };