+++ src/steps/application/index.test.ts
@@ -1,5 +1,4 @@
 import { executeStepWithDependencies } from '@jupiterone/integration-sdk-testing';
-
 import { buildStepTestConfigForStep } from '../../../test/config';
 import { Recording, setupGoogleCloudRecording } from '../../../test/recording';
 import { Steps } from '../constants';
@@ -10,13 +9,13 @@ afterEach(async () => {
   await recording.stop();
 });
 
-test('fetch-account', async () => {
+test('fetch-web-apps', async () => {
   recording = setupGoogleCloudRecording({
     directory: __dirname,
-    name: 'fetch-account',
+    name: 'fetch-web-apps',
   });
 
-  const stepConfig = buildStepTestConfigForStep(Steps.ACCOUNT);
+  const stepConfig = buildStepTestConfigForStep(Steps.WEB_APPS);
   const stepResult = await executeStepWithDependencies(stepConfig);
   expect(stepResult).toMatchStepMetadata(stepConfig);
 });