+++ CHANGELOG.md
@@ -7,3 +7,17 @@ and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+### Added
+
+- Ingest new entities
+  - `google_firebase_account`
+  - `google_firebase_project`
+  - `google_firebase_user`
+  - `google_firebase_auth_user`
+  - `google_firebase_webapp`
+- Build new relationships
+  - `google_firebase_account_has_project`
+  - `google_firebase_project_has_user`
+  - `google_firebase_project_has_auth_user`
+  - `google_firebase_project_has_webapp`